chore: Move `ContinueMode` and `LogPrefix` to `turborepo-types`
