+++ crates/turborepo-lib/src/opts.rs
@@ -9,10 +9,12 @@ use turborepo_cache::{CacheOpts, RemoteCacheOpts};
 // Re-export ScopeOpts from turborepo-scope to avoid duplication
 pub use turborepo_scope::ScopeOpts;
 use turborepo_task_id::{TaskId, TaskName};
-use turborepo_types::{DryRunMode, EnvMode, LogOrder, OutputLogsMode, UIMode};
+use turborepo_types::{
+    ContinueMode, DryRunMode, EnvMode, LogOrder, LogPrefix, OutputLogsMode, UIMode,
+};
 
 use crate::{
-    cli::{Command, ContinueMode, ExecutionArgs, LogPrefix, RunArgs},
+    cli::{Command, ExecutionArgs, RunArgs},
     config::{ConfigurationOptions, CONFIG_FILE},
     turbo_json::FutureFlags,
     Args,
@@ -557,12 +559,12 @@ mod test {
     use turbopath::AbsoluteSystemPathBuf;
     use turborepo_cache::{CacheActions, CacheConfig, CacheOpts};
     use turborepo_task_id::TaskId;
-    use turborepo_types::{DryRunMode, UIMode};
+    use turborepo_types::{ContinueMode, DryRunMode, UIMode};
     use turborepo_ui::ColorConfig;
 
     use super::{APIClientOpts, RepoOpts, RunOpts, TaskArgs};
     use crate::{
-        cli::{Command, ContinueMode, RunArgs},
+        cli::{Command, RunArgs},
         commands::CommandBase,
         config::{ConfigurationOptions, CONFIG_FILE},
         opts::{Opts, RunCacheOpts, ScopeOpts, TuiOpts},
